Organize capability definitions by convention
