+++ capabilities/assert/location_test.go
