+++ capabilities/pdp/accept.go
@@ -1,0 +1,40 @@
+package pdp
+
+import (
+	"github.com/filecoin-project/go-data-segment/merkletree"
+	"github.com/ipld/go-ipld-prime/datamodel"
+	"github.com/storacha/go-libstoracha/capabilities/types"
+	"github.com/storacha/go-libstoracha/piece/piece"
+	"github.com/storacha/go-ucanto/core/ipld"
+	"github.com/storacha/go-ucanto/core/schema"
+	"github.com/storacha/go-ucanto/validator"
+)
+
+const AcceptAbility = "pdp/accept"
+
+type AcceptCaveats struct {
+	Piece piece.PieceLink
+}
+
+func (ac AcceptCaveats) ToIPLD() (datamodel.Node, error) {
+	return ipld.WrapWithRecovery(&ac, AcceptCaveatsType(), types.Converters...)
+}
+
+var AcceptCaveatsReader = schema.Struct[AcceptCaveats](AcceptCaveatsType(), nil, types.Converters...)
+
+var Accept = validator.NewCapability(
+	AcceptAbility,
+	schema.DIDString(),
+	AcceptCaveatsReader,
+	validator.DefaultDerives,
+)
+
+type AcceptOk struct {
+	Aggregate      piece.PieceLink
+	InclusionProof merkletree.ProofData
+	Piece          piece.PieceLink
+}
+
+func (ao AcceptOk) ToIPLD() (datamodel.Node, error) {
+	return ipld.WrapWithRecovery(&ao, AcceptOkType(), types.Converters...)
+}
